task1

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned int p = 1, p1 = 0;
+	unsigned long int p = 1, p1 = 0;
 
 	if (n == 0 || n == 1)
 		_putchar(n + '0');
